% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAGS.LASSO-package.R
\docType{package}
\name{SMAGS.LASSO-package}
\alias{SMAGS.LASSO}
\alias{SMAGS.LASSO-package}
\title{SMAGS.LASSO: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
